migrated from patch-tools-example to partial-update-example project; used migrated partial-update-core(ex. patch-tools-core) library;
