+++ README.md
@@ -1,4 +1,4 @@
-# Patch Tools Example project
+# Partial Update library usage Example project
 
-This project is the example of usage Patch Tools library for HTTP PATCH
+This project is the example of usage Partial Update library for HTTP PATCH
 method logic implementation in Spring Boot application.
