+++ src/main/java/in/multitools/partial/update/example/PartialUpdateExampleApplication.java
@@ -1,0 +1,13 @@
+package in.multitools.partial.update.example;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class PartialUpdateExampleApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(PartialUpdateExampleApplication.class, args);
+	}
+
+}
