+++ src/main/java/in/multitools/partial/update/example/dto/UserDeleteDto.java
@@ -1,0 +1,4 @@
+package in.multitools.partial.update.example.dto;
+
+public class UserDeleteDto extends UserGetDto {
+}
