+++ src/main/java/in/multitools/partial/update/example/dto/UserPatchDto.java
@@ -1,0 +1,7 @@
+package in.multitools.partial.update.example.dto;
+
+import partial.update.annotation.ChangeLogger;
+
+@ChangeLogger
+public class UserPatchDto extends UserUpdateDto {
+}
