+++ src/main/java/in/multitools/partial/update/example/dto/UserUpdateDto.java
@@ -1,0 +1,4 @@
+package in.multitools.partial.update.example.dto;
+
+public class UserUpdateDto extends UserCreateDto {
+}
