+++ src/main/java/in/multitools/partial/update/example/repository/UserRepository.java
@@ -1,0 +1,7 @@
+package in.multitools.partial.update.example.repository;
+
+import in.multitools.partial.update.example.entity.User;
+import org.springframework.data.repository.CrudRepository;
+
+public interface UserRepository extends CrudRepository<User, Long> {
+}
