--- src/main/java/in/multitools/patch/tools/example/PatchToolsExampleApplication.java
@@ -1,13 +1,0 @@
-package in.multitools.patch.tools.example;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class PatchToolsExampleApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(PatchToolsExampleApplication.class, args);
-	}
-
-}
