--- src/main/java/in/multitools/patch/tools/example/dto/UserCreateDto.java
@@ -1,16 +1,0 @@
-package in.multitools.patch.tools.example.dto;
-
-import lombok.Data;
-
-import java.time.LocalDate;
-
-@Data
-public class UserCreateDto {
-
-	private String login;
-	private String firstName;
-	private String lastName;
-	private LocalDate birthDate;
-	private Boolean active;
-
-}
