--- src/main/java/in/multitools/patch/tools/example/dto/UserGetDto.java
@@ -1,17 +1,0 @@
-package in.multitools.patch.tools.example.dto;
-
-import lombok.Data;
-
-import java.time.LocalDate;
-
-@Data
-public class UserGetDto {
-
-	private Long id;
-	private String login;
-	private String firstName;
-	private String lastName;
-	private LocalDate birthDate;
-	private Boolean active;
-
-}
