--- src/main/java/in/multitools/patch/tools/example/dto/UserPatchDto.java
@@ -1,7 +1,0 @@
-package in.multitools.patch.tools.example.dto;
-
-import patch.tools.annotation.ChangeLogger;
-
-@ChangeLogger
-public class UserPatchDto extends UserUpdateDto {
-}
