--- src/main/java/in/multitools/patch/tools/example/dto/UserUpdateDto.java
@@ -1,4 +1,0 @@
-package in.multitools.patch.tools.example.dto;
-
-public class UserUpdateDto extends UserCreateDto {
-}
