--- src/main/java/in/multitools/patch/tools/example/entity/User.java
@@ -1,28 +1,0 @@
-package in.multitools.patch.tools.example.entity;
-
-import lombok.Data;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.Table;
-import java.time.LocalDate;
-
-@Data
-@Entity
-@Table(name = "user")
-public class User {
-
-	@Id()
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
-
-	private String login;
-	private String firstName;
-	private String lastName;
-	private LocalDate birthDate;
-	private boolean active;
-
-}
