--- src/main/java/in/multitools/patch/tools/example/model/UserModel.java
@@ -1,19 +1,0 @@
-package in.multitools.patch.tools.example.model;
-
-import lombok.Data;
-import patch.tools.annotation.ChangeLogger;
-
-import java.time.LocalDate;
-
-@Data
-@ChangeLogger
-public class UserModel {
-
-	private Long id;
-	private String login;
-	private String firstName;
-	private String lastName;
-	private LocalDate birthDate;
-	private boolean active;
-
-}
