--- src/main/java/in/multitools/patch/tools/example/repository/UserRepository.java
@@ -1,7 +1,0 @@
-package in.multitools.patch.tools.example.repository;
-
-import in.multitools.patch.tools.example.entity.User;
-import org.springframework.data.repository.CrudRepository;
-
-public interface UserRepository extends CrudRepository<User, Long> {
-}
