+++ src/test/java/in/multitools/partial/update/example/PartialUpdateExampleApplicationTests.java
@@ -1,0 +1,13 @@
+package in.multitools.partial.update.example;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class PartialUpdateExampleApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
