--- src/test/java/in/multitools/patch/tools/example/PatchToolsExampleApplicationTests.java
@@ -1,13 +1,0 @@
-package in.multitools.patch.tools.example;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class PatchToolsExampleApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
